feat: implement the empty operation
This will implement the `empty` action available in the console(`rm
--recursive` in the cli). I want the behavior to be drawn from the
behavior of the cli; the function will return a list of keys that
were deleted. I'm also thinking about adding a `quiet` option that
simply returns the keys that failed to delete, but probably not for
the first version.