feat: add listObject implementation (#11)
This commit add the `listObject` operation as well as adding useful
scripts to run tests. The test for `listObject` is currently ignored due
to inconsistencies between the s3 api and the minio api that we use for
tests. As-is the code works and has been tested against an actual s3
bucket, but the test doesn't work; I want to figure out a way to have it
both ways later on.

Co-authored-by: Luca Casonato <luca.casonato@antipy.com>